refactor: split correlations.ts into domain-focused modules
Split the 1496-line correlations service into 8 focused modules:
- types.ts: All interfaces/types
- utils.ts: Statistical helpers (mean, stddev, pearson, chi-squared)
- baseline.ts: getBaseline
- hrv-activities.ts: getHrvActivitiesCorrelation
- activity-impact.ts: getActivityImpact
- event-probability.ts: getEventProbability
- generic.ts: getGenericCorrelation + helpers
- index.ts: Barrel re-export
